Fix memory requirements for draw_clear tests
The memory requirements for the draw_clear subtests of
the conditional_rendering group required the memory
bound for the conditional rendering predicate buffer
to be both host visible and device local.

It's not a requirement that all implementation support
this combinations so all the draw_clear tests were
skipped unnecessarily.

Since the device local requirement was unnecessary,
this CL just removes it.

Affects:
dEQP-VK.conditional_rendering.draw_clear.*

Change-Id: I10d289a7666a97d665707d5e49ec7c8846d2997d
Components: Vulkan